resolve import cycle
